requirements.txt
groq
langchain-core 
langchain-community
langchain-groq
pyairtable
python-dotenv
